requirements_pip.txt
avocado-framework==109.0
